+++ scripts/set-flavour.js
@@ -1,13 +1,13 @@
 import { readdirSync, readFileSync, writeFileSync } from 'fs';
 
 if (process.argv.length <= 2) {
-  console.error('Please provide a flavour (DEV, STANDALONE, STEAM, STEAM_CN)');
+  console.error('Please provide a flavour (DEV, STANDALONE, STEAM)');
   process.exit(1);
 }
 
 const flavour = process.argv[2].toUpperCase();
-if (!['DEV', 'STANDALONE', 'STEAM', 'STEAM_CN'].includes(flavour)) {
-  console.error('Provided flavour is not valid (DEV, STANDALONE, STEAM, STEAM_CN)');
+if (!['DEV', 'STANDALONE', 'STEAM'].includes(flavour)) {
+  console.error('Provided flavour is not valid (DEV, STANDALONE, STEAM)');
   process.exit(1);
 }
 
@@ -20,8 +20,6 @@ let appKey = (() => {
       return 'steam.overlay.2538150-STANDALONE';
     case 'STEAM':
       return 'steam.overlay.2538150-STEAM';
-    case 'STEAM_CN':
-      return 'steam.overlay.2538150-STEAM';
     default:
       console.warn('COULD NOT DETERMINE APP KEY FROM FLAVOUR');
       return 'steam.overlay.2538150-DEV';
@@ -33,7 +31,7 @@ let appKey = (() => {
   const path = 'src-ui/build.ts';
   let uiFlavour = readFileSync(path).toString();
   uiFlavour = uiFlavour.replaceAll(
-    /export const FLAVOUR: BuildFlavour = '(DEV|STANDALONE|STEAM|STEAM_CN)';/g,
+    /export const FLAVOUR: BuildFlavour = '(DEV|STANDALONE|STEAM)';/g,
     `export const FLAVOUR: BuildFlavour = '${flavour}';`
   );
   writeFileSync(path, uiFlavour);
@@ -45,7 +43,7 @@ let appKey = (() => {
   const path = 'src-core/src/flavour.rs';
   let coreFlavour = readFileSync(path).toString();
   coreFlavour = coreFlavour.replaceAll(
-    /pub const BUILD_FLAVOUR: BuildFlavour = BuildFlavour::(Dev|Standalone|Steam|SteamCn);/g,
+    /pub const BUILD_FLAVOUR: BuildFlavour = BuildFlavour::(Dev|Standalone|Steam);/g,
     `pub const BUILD_FLAVOUR: BuildFlavour = BuildFlavour::${flavour
       .split('_')
       .map((t) => t.toUpperCase().charAt(0) + t.toLowerCase().substring(1))