+++ src-core/src/flavour.rs
@@ -5,7 +5,6 @@ pub enum BuildFlavour {
     Dev,
     Standalone,
     Steam,
-    SteamCn,
 }
 
 pub const BUILD_FLAVOUR: BuildFlavour = BuildFlavour::Dev;